+++ src/layouts/sidebar/components/MenuItem.vue
@@ -5,7 +5,7 @@
   >
     <!-- <SvgIcon :icon-class="showRoute.meta ? showRoute.meta.icon || 'el-icon-menu' : 'el-icon-menu'" /> -->
     <el-icon>
-      <component :is="showRoute.meta ? showRoute.meta.icon || MenuIcon : MenuIcon" />
+      <component :is="showRoute.meta ? showRoute.meta.icon || OperationIcon : OperationIcon" />
     </el-icon>
     <template #title>
       <span>{{ showRoute.meta ? showRoute.meta.title : showRoute.name }}</span>
@@ -22,7 +22,7 @@ import { isExternal } from "../../utils";
 import path from "path";
 import { defineComponent } from "vue";
 import { useRouter, useRoute } from "vue-router";
-import { Menu as MenuIcon } from "@element-plus/icons";
+import { Operation as OperationIcon } from "@element-plus/icons";
 export default defineComponent({
   name: "MenuItem",
   props: {
@@ -73,7 +73,7 @@ export default defineComponent({
     return {
       generatorPath,
       handleClick,
-      MenuIcon,
+      OperationIcon,
     };
   },
 });