+++ src/views/form/advance-form.vue
@@ -8,7 +8,7 @@
         <div class="flex justify-between padding-sm">
           <el-link :underline="false">添加学校</el-link>
           <el-button
-            size="small"
+            size="default"
             type="primary"
           >立即添加</el-button>
         </div>
@@ -26,7 +26,7 @@
           v-model="baseInfoModel"
           label-width="80px"
           class="form-wrapper"
-          size="small"
+          size="default"
           label-position="right"
         >
           <el-row>
@@ -54,7 +54,7 @@
               <el-form-item label="加盟方式">
                 <el-radio-group
                   v-model="baseInfoModel.joinType"
-                  size="small"
+                  size="default"
                 >
                   <el-radio-button label="普通" />
                   <el-radio-button label="独家" />
@@ -98,7 +98,7 @@
           v-model="baseInfoModel"
           label-width="80px"
           class="form-wrapper"
-          size="small"
+          size="default"
           label-position="right"
         >
           <el-row>
@@ -124,7 +124,7 @@
               <el-form-item label="学生数量">
                 <el-input-number
                   v-model="baseInfoModel.studentNum"
-                  size="small"
+                  size="default"
                 />
               </el-form-item>
             </el-col>
@@ -149,7 +149,7 @@
       >
         <div class="flex justify-end padding-sm">
           <el-button
-            size="small"
+            size="default"
             type="primary"
           >立即添加</el-button>
         </div>