+++ src/views/list/table-with-search.vue
@@ -1,17 +1,19 @@
 <template>
   <div class="main-container">
-    <TableHeader
-      :can-collapsed="
+    <TableBody>
+      <template #header>
+        <TableHeader
+          :can-collapsed="
         likeSearchModel.conditionItems &&
         likeSearchModel.conditionItems.length !== 0
       "
-      :search-model="likeSearchModel.conditionItems"
-      :default-collapsed-state="true"
-      title="数据筛选"
-      @doSearch="doSearch"
-      @resetSearch="resetSearch"
-    />
-    <TableBody>
+          :search-model="likeSearchModel.conditionItems"
+          :default-collapsed-state="true"
+          title="表格查询"
+          @doSearch="doSearch"
+          @resetSearch="resetSearch"
+        />
+      </template>
       <template #tableConfig>
         <TableConfig
           v-model:border="tableConfig.border"
@@ -19,11 +21,17 @@
           @refresh="doRefresh"
         >
           <template #actions>
-            <el-button type="primary" size="small" icon="PlusIcon"
-              >添加
+            <el-button
+              type="primary"
+              size="small"
+              icon="PlusIcon"
+            >添加
             </el-button>
-            <el-button type="danger" size="small" icon="DeleteIcon"
-              >删除
+            <el-button
+              type="danger"
+              size="small"
+              icon="DeleteIcon"
+            >删除
             </el-button>
           </template>
         </TableConfig>
@@ -38,13 +46,24 @@
           :border="tableConfig.border"
           :height="tableConfig.height"
         >
-          <el-table-column align="center" label="序号" width="80">
+          <el-table-column
+            align="center"
+            label="序号"
+            width="80"
+          >
             <template v-slot="scope">
               {{ scope.$index + 1 }}
             </template>
           </el-table-column>
-          <el-table-column align="center" label="名称" prop="nickName" />
-          <el-table-column align="center" label="头像">
+          <el-table-column
+            align="center"
+            label="名称"
+            prop="nickName"
+          />
+          <el-table-column
+            align="center"
+            label="头像"
+          >
             <template v-slot="scope">
               <div class="avatar-container">
                 <el-image
@@ -60,7 +79,11 @@
               </div>
             </template>
           </el-table-column>
-          <el-table-column align="center" label="性别" prop="gender">
+          <el-table-column
+            align="center"
+            label="性别"
+            prop="gender"
+          >
             <template v-slot="scope">
               <div class="gender-container flex justify-center align-center">
                 <img
@@ -75,7 +98,10 @@
               </div>
             </template>
           </el-table-column>
-          <el-table-column align="center" label="状态">
+          <el-table-column
+            align="center"
+            label="状态"
+          >
             <template v-slot="scope">
               <el-switch
                 v-model="scope.row.status"
@@ -84,7 +110,11 @@
               />
             </template>
           </el-table-column>
-          <el-table-column align="center" label="地址" prop="address" />
+          <el-table-column
+            align="center"
+            label="地址"
+            prop="address"
+          />
           <el-table-column
             align="center"
             label="上次登录时间"